Add functions for collect exam statistics.

+++ src/main/java/com/mumu/Online/Exam/System/model/entity/ExamLogin.java
@@ -33,7 +33,7 @@ public class ExamLogin extends BaseModel {
 
     private String blankAnswers;
 
-    private float score;
+    private Double score;
 
     private Date loginDate;
